Add listen function.
