--- plugins/mention.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from slackbot.bot import listen_to
-import random
-
-
-@listen_to('リッスン')
-def listen_func(message):
-    message.react('atsumori')
-    if random.random() < 0.5:
-        message.send('失礼しました。熱盛と出てしまいました。')
-    else:
-        message.send('あっ…熱盛が出てしまい…ました失礼しました。')
